// src/shared/interfaces/ISearchResult.ts
export interface ISearchResult {
    id: string;
    ano: number;
    anoModelo: number;
    chassi: string;
    codigoRetorno: number;
    codigoSituacao: number;
    cor: string;
    data: string;
    dataAtualizacaoAlarme: string;
    dataAtualizacaoCaracteristicasVeiculo: string;
    dataAtualizacaoRouboFurto: string;
    extra: Object;
    marca: string;
    mensagemRetorno: string;
    modelo: string;
    municipio: string;
    placa: string;
    situacao: string;
    uf: string;
}